Remove GenericSelectTarget
